+++ hyrisecockpit/frontend/tests/e2e/specs/views/abstractTests.ts
