+++ hyrisecockpit/frontend/tests/e2e/specs/views/helpers.ts
